dev: add action button (missing icon version)
