+++ src/components/ActionButton/ActionButton.scss
@@ -1,0 +1,114 @@
+// primary
+$action-button-primary-color-font: $surgical-500;
+$action-button-primary-color-background: $bone;
+$action-button-primary-color-border: $surgical-500;
+// primary:hover
+$action-button-primary-hover-color-background: $surgical-100;
+$action-button-primary-hover-color-box-shadow: $surgical-400;
+$action-button-primary-hover-color-border: $surgical-100;
+// primary:disabled
+$action-button-primary-disabled-color-background: $ash;
+$action-button-primary-disabled-color-font: $waiting-room;
+
+// secondary
+$action-button-secondary-color-font: $surgical-500;
+$action-button-secondary-color-background: transparent;
+$action-button-secondary-color-border: $surgical-500;
+// secondary:hover
+$action-button-secondary-hover-color-border: $surgical-100;
+$action-button-secondary-hover-color-box-shadow: $surgical-400;
+$action-button-secondary-hover-color-background: $surgical-100;
+// secondary:disabled
+$action-button-secondary-disabled-color-border: $waiting-room;
+$action-button-secondary-disabled-color-font: $waiting-room;
+
+// tertiary
+$action-button-tertiary-color-font: $surgical-500;
+$action-button-tertiary-color-background: transparent;
+// tertiary:hover
+$action-button-tertiary-hover-color-border: $surgical-100;
+$action-button-tertiary-hover-color-box-shadow: $surgical-400;
+$action-button-tertiary-hover-color-background: $surgical-100;
+// tertiary:disable
+$action-button-tertiary-disabled-color-border: $waiting-room;
+$action-button-tertiary-disabled-color-font: $waiting-room;
+
+.is-action {
+  background-color: $action-button-primary-color-background;
+  border: 0.125rem solid $action-button-primary-color-border;
+  color: $action-button-primary-color-font;
+  @include button-variant-borders(
+    $action-button-primary-hover-color-border,
+    $action-button-primary-hover-color-box-shadow
+  );
+  &:disabled,
+  &.is-disabled {
+    color: $action-button-primary-disabled-color-font;
+    background-color: $action-button-primary-disabled-color-background;
+  }
+  &:hover,
+  &.is-hovered {
+    background-color: $action-button-primary-hover-color-background;
+    border-color: $action-button-primary-hover-color-border;
+  }
+  &:active,
+  &.is-active {
+    background-color: $action-button-primary-hover-color-background;
+    border-color: $action-button-primary-hover-color-border;
+  }
+  &:focus,
+  &.is-focused {
+    background-color: $action-button-primary-hover-color-background;
+    border-color: $action-button-primary-hover-color-border;
+  }
+  &.is-secondary {
+    @include button-variant-borders(
+      $action-button-secondary-hover-color-border,
+      $action-button-secondary-hover-color-box-shadow
+    );
+    background-color: $action-button-secondary-color-background;
+    border: 0.125rem solid transparent;
+    color: $action-button-secondary-color-font;
+    &:disabled,
+    &.is-disabled {
+      color: $action-button-secondary-disabled-color-font;
+      border-color: $action-button-secondary-disabled-color-border;
+    }
+    &:hover,
+    &.is-hovered {
+      background-color: $action-button-secondary-hover-color-background;
+      border: 0.125rem solid $action-button-secondary-hover-color-border;
+    }
+    &:active,
+    &.is-active {
+      background-color: $action-button-secondary-hover-color-background;
+      border: 0.125rem solid $action-button-secondary-hover-color-border;
+    }
+    &:focus,
+    &.is-focused {
+      background-color: $action-button-secondary-hover-color-background;
+      border: 0.125rem solid $action-button-secondary-hover-color-border;
+    }
+  }
+  &.is-tertiary {
+    border: 0;
+    background-color: $action-button-tertiary-color-background;
+    color: $action-button-tertiary-color-font;
+    &:disabled,
+    &.is-disabled {
+      border: 0;
+    }
+    &:hover,
+    &.is-hovered &:active,
+    &.is-active,
+    &:focus,
+    &.is-focused {
+      &:before {
+        border: 0;
+      }
+      &:after {
+        border: 0;
+      }
+    }
+  }
+}
