+++ src/components/ActionButton/index.ts
@@ -1,0 +1,1 @@
+export {ActionButton} from './ActionButton';
