+++ src/components/ConfirmationButton/ConfirmationButton.scss
@@ -1,6 +1,3 @@
-@import '../../styling/palette';
-@import '../../styling/components/button';
-
 // primary
 $confirmation-button-primary-color-font: $bone;
 $confirmation-button-primary-color-background: $strangulation-400;
@@ -26,7 +23,7 @@ $confirmation-button-secondary-disabled-color-font: $waiting-room;
 
 // tertiary
 $confirmation-button-tertiary-color-font: $strangulation-500;
-$confirmation-button-tertiary-color-background: $bone;
+$confirmation-button-tertiary-color-background: transparent;
 // tertiary:hover
 $confirmation-button-tertiary-hover-color-border: $strangulation-100;
 $confirmation-button-tertiary-hover-color-background: $strangulation-100;
