+++ src/index.ts
@@ -1,3 +1,4 @@
+export {ActionButton} from './components/ActionButton';
 export {StartButton} from './components/StartButton';
 export {ConfirmationButton} from './components/ConfirmationButton';
 export {Icon} from './components/Icons';
