+++ src/styling/helsenorge.scss
@@ -8,6 +8,7 @@
 
 @import 'components/button';
 
+@import '../components/ActionButton/ActionButton';
 @import '../components/CommonButton/CommonButton';
 @import '../components/ConfirmationButton/ConfirmationButton';
 @import '../components/StartButton/StartButton';
