--- stories/Buttons/CommonButton.stories.tsx
@@ -1,49 +1,0 @@
-import React from 'react';
-
-import {text, select, boolean, withKnobs} from '@storybook/addon-knobs';
-import {storiesOf} from '@storybook/react';
-
-import Stacker from '../../.storybook/Stacker/Stacker';
-
-const stories = storiesOf('Buttons', module);
-stories.addDecorator(withKnobs);
-
-import {CommonButton} from '../../src/components/CommonButton';
-
-const buttonVariants = {
-  primary: null,
-  secondary: 'secondary',
-  tertiary: 'tertiary'
-}
-
-stories.add('CommonButton', (): JSX.Element => (
-    <CommonButton
-      variant={select('variant', buttonVariants, buttonVariants.primary)}
-      isLoading={boolean('isLoading', false)}
-      disabled={boolean('disabled', false)}
-      >
-      {text('text', 'CommonButton')}
-    </CommonButton>
-));
-
-stories.add('CommonButton (all variants)', (): JSX.Element => (
-  <Stacker vertical={boolean('vertically stacking', false)}>
-    <CommonButton
-      disabled={boolean('disabled', false)}
-      >
-      Primary
-    </CommonButton>
-    <CommonButton
-      variant="secondary"
-      disabled={boolean('disabled', false)}
-      >
-      Secondary
-    </CommonButton>
-    <CommonButton
-      variant="tertiary"
-      disabled={boolean('disabled', false)}
-      >
-      Tertiary
-    </CommonButton>
-  </Stacker>
-));
